backend/requirements.txt
# Core Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.6.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
alembic>=1.14.0

# Redis
redis>=5.2.0

# LLM Providers
openai>=1.58.0
anthropic>=0.40.0
google-generativeai>=0.8.3

# HTTP Client
httpx>=0.28.0

# Authentication
PyJWT>=2.8.0

# Utilities
python-dotenv>=1.0.1
python-multipart>=0.0.20
